Make Datepicker and Timepicker classes private in their widgets
