+++ library/src/main/java/com/vijay/jsonwizard/fragments/JsonFormFragment.java
@@ -40,8 +40,6 @@ import com.vijay.jsonwizard.expressions.JsonExpressionResolver;
 import com.vijay.jsonwizard.i18n.JsonFormBundle;
 import com.vijay.jsonwizard.interfaces.CommonListener;
 import com.vijay.jsonwizard.interfaces.JsonApi;
-import com.vijay.jsonwizard.listeners.DatePickerListener;
-import com.vijay.jsonwizard.listeners.TimePickerListener;
 import com.vijay.jsonwizard.maps.MapsUtils;
 import com.vijay.jsonwizard.mvp.MvpFragment;
 import com.vijay.jsonwizard.presenters.JsonFormFragmentPresenter;
